Move SQLite factor host impl to own module.
